pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "shamefile"
version = "0.1.0"
description = "A cli tool to enforce documentation for code suppressions"
authors = [
    {name = "Bartlomiej Flis", email = "bartekdawidflis@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pre-commit>=2.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pyyaml>=6.0",
    "ruff>=0.15",
]

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # trailing comma style — conflicts with ruff format
    "ISC001",  # implicit string concat — conflicts with ruff format
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",    # assert is fine in tests
    "ANN",     # no type annotations needed for tests
    "D",       # no docstring enforcement for tests
    "INP001",  # no __init__.py required for test dirs
    "S603",    # subprocess calls are intentional in integration tests
    "S607",    # partial executable path (git, cargo) is fine in tests
]

[tool.maturin]
features = [] # No python bindings features needed for pure binary
bindings = "bin" # Ensure we build the binary, not a lib
strip = true 
